flask/requirements.txt
Flask
flask-cors
python-dotenv
Pillow
requests
google-generativeai